requirements.txt
asgiref==3.7.2
coverage==7.4.0
Django==5.0.1
django-mysql==4.12.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
PyMySQL==1.1.0
pytest==7.4.4
pytest-django==4.7.0
python-decouple==3.8
sqlparse==0.4.4
